code/portal/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
pydantic==1.10.8
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
boto3==1.35.74
botocore==1.35.74
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.1
openpyxl==3.1.2
numpy==1.26.0
bcrypt==4.0.1